requirements.txt
django>=4.0.5
django-allauth==0.43.0
django-crispy-forms==1.11.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
pillow==9.1.1
humanfriendly==10.0
channels==3.0.5
channels-redis==3.4.1
celery==5.2.7
djangorestframework==3.13.1